--- src/main/java/goorm/bus/BusApplication.java
@@ -1,23 +1,0 @@
-package goorm.bus;
-
-import io.github.cdimascio.dotenv.Dotenv;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
-@EnableJpaAuditing
-@EnableScheduling
-public class BusApplication {
-
-	public static void main(String[] args) {
-		// env
-		Dotenv dotenv = Dotenv.load();
-		dotenv.entries().forEach(entry -> System.setProperty(entry.getKey(), entry.getValue()));
-
-		SpringApplication.run(BusApplication.class, args);
-	}
-
-}
