+++ src/main/java/goorm/domain/record/dto/response/NoteResponse.java
@@ -1,6 +1,6 @@
-package goorm.bus.record.dto.response;
+package goorm.domain.record.dto.response;
 
-import goorm.bus.record.entity.Note;
+import goorm.domain.record.entity.Note;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -8,7 +8,7 @@ import lombok.Builder;
 
 
 @Builder
-public record NoteSaveResponse(
+public record NoteResponse(
 
         @NotNull
         @Schema(description = "게시물 ID", example = "1")
@@ -29,24 +29,37 @@ public record NoteSaveResponse(
         @Schema(description = "넘겨준 시간", example = "23:00")
         String time,
 
+
+        @NotBlank
+        @Schema(description = "알람활성화", example = "true or false")
+        boolean alarm,
+
+        @NotBlank
+        @Schema(description = "즐겨찾기", example = "true or false")
+        boolean favorite,
+
         @NotBlank
         @Schema(description = "노선 즉 버스ID ", example = "356")
         String notionId,
 
         @NotBlank
         @Schema(description = "정류장 ID", example = "54686")
-        String stationId
+        String stationId,
+
+        @NotBlank
+        @Schema(description = "최대빈도수", example = "2")
+        int frequency
 
 ) {
-    public static NoteSaveResponse of(Note note) {
-        return NoteSaveResponse.builder()
+    public static NoteResponse of(Note note) {
+        return NoteResponse.builder()
                 .id(note.getId())
                 .departure(note.getDeparture())
                 .destination(note.getDestination())
                 .station(note.getStation())
                 .time(note.getTime())
-                .stationId(note.getStationId())
-                .notionId(note.getNotionId())
+                .alarm(note.isAlarm())
+                .frequency(note.getFrequency())
                 .build();
     }
 }
